# lab_wizard/wizard/frontend/src/lib/api/openapi.json
{
 "components": {
  "schemas": {
...
   "FileSettings-Input": {
    "additionalProperties": false,
    "description": "Where each run's folder of CSV and YAML files goes, and what it holds.",
    "properties": {
     "path": {
      "default": "{date}/{procedure}_{device}_{time}",
      "description": "Where each run goes under the root, from filter keys: {date} {time} {procedure} {device} {device.<property>} {operator} {run.<metadata>} {param.<path>} {run_id}",
      "title": "Path",
      "type": "string"
     },
     "plot_png": {
      "default": true,
      "description": "Also save the run's default plot as plot.png",
      "title": "Plot Png",
      "type": "boolean"
     },
     "root": {
      "default": "",
      "description": "Folder to save runs under; empty for the workspace's data/files. A relative path is relative to the workspace.",
      "title": "Root",
      "type": "string"
     }
    },
    "title": "FileSettings",
    "type": "object"
   },
   "FileSettings-Output": {
    "additionalProperties": false,
    "description": "Where each run's folder of CSV and YAML files goes, and what it holds.",
    "properties": {
     "path": {
      "default": "{date}/{procedure}_{device}_{time}",
      "description": "Where each run goes under the root, from filter keys: {date} {time} {procedure} {device} {device.<property>} {operator} {run.<metadata>} {param.<path>} {run_id}",
      "title": "Path",
      "type": "string"
     },
     "plot_png": {
      "default": true,
      "description": "Also save the run's default plot as plot.png",
      "title": "Plot Png",
      "type": "boolean"
     },
     "root": {
      "default": "",
      "description": "Folder to save runs under; empty for the workspace's data/files. A relative path is relative to the workspace.",
      "title": "Root",
      "type": "string"
     }
    },
    "required": [
     "root",
     "path",
     "plot_png"
    ],
    "title": "FileSettings",
    "type": "object"
   },
   "FileSettingsView": {
    "properties": {
     "example": {
      "title": "Example",
      "type": "string"
     },
     "files": {
      "$ref": "#/components/schemas/FileSettings-Output"
     },
     "folder": {
      "title": "Folder",
      "type": "string"
     },
     "keys": {
      "items": {
       "type": "string"
      },
      "title": "Keys",
      "type": "array"
     },
     "problems": {
      "items": {
       "$ref": "#/components/schemas/TemplateProblem"
      },
      "title": "Problems",
      "type": "array"
     }
    },
    "required": [
     "example",
     "problems",
     "files",
     "folder",
     "keys"
    ],
    "title": "FileSettingsView",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "TemplateCheck": {
    "properties": {
     "example": {
      "title": "Example",
      "type": "string"
     },
     "problems": {
      "items": {
       "$ref": "#/components/schemas/TemplateProblem"
      },
      "title": "Problems",
      "type": "array"
     }
    },
    "required": [
     "example",
     "problems"
    ],
    "title": "TemplateCheck",
    "type": "object"
   },
   "TemplateProblem": {
    "description": "Something wrong with a folder template: an error blocks saving, a warning does not.",
    "properties": {
     "key": {
      "title": "Key",
      "type": "string"
     },
     "level": {
      "enum": [
       "error",
       "warning"
      ],
      "title": "Level",
      "type": "string"
     },
     "message": {
      "title": "Message",
      "type": "string"
     }
    },
    "required": [
     "level",
     "key",
     "message"
    ],
    "title": "TemplateProblem",
    "type": "object"
   },
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
   "_TemplateBody": {
    "properties": {
     "path": {
      "title": "Path",
      "type": "string"
     }
    },
    "required": [
     "path"
    ],
    "title": "_TemplateBody",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/settings/files": {
   "get": {
    "description": "How runs of projects with file saving on are laid out on disk.",
    "operationId": "api_settings_files_api_settings_files_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FileSettingsView"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Api Settings Files"
   },
   "put": {
    "operationId": "api_settings_files_save_api_settings_files_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FileSettings-Input"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FileSettingsView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Settings Files Save"
   }
  },
  "/api/settings/files/check": {
   "post": {
    "description": "What a folder template would name the latest run's folder, and what is wrong with it.",
    "operationId": "api_settings_files_check_api_settings_files_check_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/_TemplateBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TemplateCheck"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Api Settings Files Check"
   }
  },
...
 }
}